add refactor send comment callback

+++ webService/src/main/java/ru/newsystems/webservice/handler/messaga/MessageExitHandler.java
@@ -1,9 +1,7 @@
 package ru.newsystems.webservice.handler.messaga;
 
 import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardRemove;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import ru.newsystems.basecore.integration.VirtaBot;
 import ru.newsystems.basecore.model.state.MessageState;
@@ -26,6 +24,7 @@ public class MessageExitHandler implements MessageHandler{
     @Override
     public boolean handleUpdate(Update update) throws TelegramApiException {
         String text = update.getMessage().getText();
+        if (text == null) return false;
         if (MessageState.getState(text).equals(MessageState.EXIT)) {
             closeReplyKeyBoard(update, bot, true);
             localRepo.remove(update.getMessage().getChatId());
